fix: revert pinning smallvec to 1.10.0
Tried to manually patch `smallvec` to a non-vulnerable version but this causes downstream dependencies to be unable to resolve a version not equal to 1.10.0 

Regardless I don't think cargo is able to resolve the _ancient_ version of `smallvec` used by `tokio-proto` anyways. We'll need to accept the risk for this vuln until we're ready to move to fred 5/6.